+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Direqt SDK for JavaScript - Change Log
 
+## v1.4.0
+
+ * Add support for webhook configuration APIs
+ * Add webhook sample
+
 ## v1.3.0
  * The Messaging API access_token is now passed as an Authorization header, rather than as a query string parameter.
  * Updated the Messaging API endpoint to canonical https://gateway.direqt.io/v3/messages